[FIX] Properly set support helpdesk team.
Formerly we were trying to create a demo helpdesk team named Support.
This do not raise any error in the server log however after review the
instance I found out that the support team was not created.

This is because there is another helpdesk team define as real data in
the system with the same name.

I jut update my demo data to relate with the support real data team and
now the support does set with the values I added in the xml file.