[FIX] Warnings in instance odoo log
When we add the functionality of Invoiceable lines we add the
hr_timesheet_reports. This were a functionality moved form version 8.0.
This functionality is really not working today but we migrate only the basic
need to run the code: doing this some warnings were leave in code.

Now is the time to fix them by updating the methods api to new api.