[ADD] Show product information in the project task
When a task has been created from a sale order is actually linked to a sale
order line which product requires to create a task. With this change the product
how was responsible to originate that task will be shown in the task extra
information tab when this one is set.