[FIX] Load language only once when installing the module
Formerly, Language was installed 2 time each time we update instance: one when
we install the language and two when we install the website languages.

Now, only run one time when the vauxoo module is installed (delete double
installation). It will not run the language installation when the module is
updated.

NOTE: I use vauxoo/mexico and vauxoo/yoytec as example to make this change