[FIX] Add customer var to the gitlab-ci file.
This var is needed because we added the posibility to build more than one image in the upload step as can be seen [here](https://git.vauxoo.com/vauxoo/gitlab_tools/blob/master/src/deployv_addon_gitlab_tools/commands/upload_image.py#L71) if that var is not defined it will use the repo name by default. In the case of Vauxoo, the repo is named instance, but the image vauxoo (plus the version).