task#7678 [FIX] Count of tasks of project take into account Orphan tasks.
Formerly all the tasks that has the tag "*Orphan*" was not taked into account
when counting the tasks related to a project. Now this criteria was removed
from the domain and all the tasks that are parent tasks are taked into account
in the counting, not taking into account the tag name.