[ADD] New template for portal quotations without default values
Add a new template that makes that the portal quotation shows the
section names but leave empty the default text that odoo set.

The old default template of odoo was deleted, since this one does not
have an active field we can set. Not was my intention to delete the
template at the beginning but is the way I found to "disable" the
template.