[FIX] Restore default odoo quotation template removed in last iteration
In a first place the idea was to disable this template, but seems this
model was not able to be disable the decision was remove it.

After review, we decide that this template will not be deleted, but,
will be re used as example for other templates and will be updated
in production as well.