[ADD] Enable l10n_mx_edi module dependency
This one was disable by a red we have in our builds that do not let us
to merge anything in instance vauxoo. The red was related to l10n_mx
certificate expired problem.

NOTE: disable base_address_extended that was enable just to avoid a
warning when we disable the l10n_mx_edi dependency.