[ADD] Vauxoo instance does not manage pads anymore.
This change can be reverted later or even activate manually in a
production database if needed installing the pad module.

NOTE: pad_project is a auto_installable module when both pad and
project is installed. So in order to avoid this auto installation
I been force to remove the pad module installation and its configuration.