[ADD] New server action that will properly compute the tasks spent hours.
This is required when migration data since the timesheet is added after
the task has been loaded and we need to compute this field since this
one is not updated later.

NOTE: Add extra new generic method named load_and_run that will let us
to load a csv action server and run the action server.