[FIX] Enable website_settings
The website.config.settings module was merged to res.config.settings, so
here we update and move the related to the res_config_settings file.

Also social_ fields from website_settings were removed from wizard
model: this ones can now be set in the company or in the website.

By now the social_ fields were set in the main_company.